'again_pull'

+++ models/user.js
@@ -25,7 +25,7 @@ module.exports = (sequelize, Sequelize) => {
                 allowNull: false,
                 defaultValue: ""
             },
-            token:{
+            token: {
                 type: Sequelize.STRING,
                 allowNull: true,
             },
@@ -61,7 +61,7 @@ module.exports = (sequelize, Sequelize) => {
                 type: Sequelize.DOUBLE(20, 6),
                 allowNull: false
             },
-            role : {
+            role: {
                 type: Sequelize.STRING,
                 alllowNull: false
             }
@@ -74,9 +74,9 @@ module.exports = (sequelize, Sequelize) => {
     );
 
     User.migrate = async () => {
-        // const count = await User.count();
+        const count = await User.count();
 
-        // if (!count) {
+        if (!count) {
             await User.destroy({ truncate: true });
             const saltRounds = 10;
             const hashedPassword = await bcrypt.hash("123456", saltRounds);
@@ -91,7 +91,7 @@ module.exports = (sequelize, Sequelize) => {
                 role: "admin",
                 commission: 0.03,
             })
-        // }
+        }
     };
 
     return User;
